+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### vNext
+### v3.3.0
 
 - Copy / Paste Duplicant settings.
 - Clone Duplicant.